+++ assets/javascripts/kitten/components/cards/contribution-card/components/index.js
@@ -1,0 +1,4 @@
+export { Amount } from './amount'
+export { Image } from './image'
+export { PillNumber } from './pill-number'
+export { Title } from './title'