+++ assets/javascripts/kitten/components/cards/contribution-card/components/pill-number.js
@@ -1,0 +1,19 @@
+import React from 'react'
+import classNames from 'classnames'
+import { PillNumberInput } from '../../../../components/form/pill-number-input'
+
+export const PillNumber = ({ value, min, max, ...props }) => {
+  return (
+    <PillNumberInput 
+      value={value}
+      min={min}
+      max={max}
+      className={classNames(
+        'k-ContributionCard__pillNumber',
+        'k-u-margin-right@s-up',
+        props.className,
+      )}
+      {...props}
+    />
+  )
+}